Move login guard to shared module
