+++ client/app/src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import {
 } from './shared/_components/redirect-to-default-route/redirect-to-default-route.component';
 import { CreateProfileGuardService } from './user-profile/create-profile-guard.service';
 
-import { LoginGuard } from './core/auth/_guards/login-guard.service';
+import { LoginGuard } from './shared/_guards/login-guard.service';
 
 const routes: Routes = [
   { path: '**', redirectTo: '/redirect' }, // redirects to '' in case no route is matched
