+++ client/app/src/app/case-management/case-management-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { LoginGuard } from '../core/auth/_guards/login-guard.service';
+import { LoginGuard } from '../shared/_guards/login-guard.service';
 import { CreateProfileGuardService } from '../user-profile/create-profile-guard.service';
 import { FormListComponent } from './form-list/form-list.component';
 import { CaseDetailsComponent } from './case-details/case-details.component';
