+++ client/app/src/app/class/reports/student-grouping-report/student-grouping-report.component.spec.ts
@@ -21,9 +21,5 @@ describe('StudentGroupingReportComponent', () => {
     fixture.detectChanges();
   });
 
-  /*
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-  */
+  it('should create');
 });
