+++ client/app/src/app/class/reports/student-progress-table/student-progress-table.component.spec.ts
@@ -20,9 +20,5 @@ describe('StudentProgressTableComponent', () => {
     fixture.detectChanges();
   });
 
-  /*
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-  */
+  it('should create');
 });
