+++ client/app/src/app/class/reports/student-subtest-report/student-subtest-report.component.spec.ts
@@ -21,7 +21,5 @@ describe('StudentSubtestReportComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+  it('should create');
 });
