+++ client/app/src/app/class/reports/task-report/task-report.component.spec.ts
@@ -21,7 +21,5 @@ describe('TaskReportComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+  it('should create');
 });
