+++ client/app/src/app/core/auth/auth-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { RegistrationComponent } from './registration/registration.component';
-import { LoginGuard } from './_guards/login-guard.service';
+import { LoginGuard } from '../../shared/_guards/login-guard.service';
 
 // @TODO Add edit-profile component.
 // import { EditProfileComponent } from './edit-profile/edit-profile.component';
