+++ client/app/src/app/core/auth/auth.module.ts
@@ -2,9 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule, MatInputModule, MatSelectModule } from '@angular/material';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { LoginGuard } from './_guards/login-guard.service';
 import { AuthRoutingModule } from './auth-routing.module';
 import { LoginComponent } from './login/login.component';
 import { RegistrationComponent } from './registration/registration.component';
@@ -19,11 +17,9 @@ import { SharedModule } from '../../shared/shared.module';
     MatSelectModule,
     FormsModule,
     ReactiveFormsModule,
-    BrowserAnimationsModule,
     AuthRoutingModule,
     SharedModule
   ],
-  providers: [LoginGuard],
   declarations: [LoginComponent, RegistrationComponent]
 })
 export class AuthModule { }
