+++ client/app/src/app/core/auth/login/login.component.spec.ts
@@ -20,7 +20,5 @@ describe('LoginComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should be created', () => {
-    expect(component).toBeTruthy();
-  });
+  it('should be created');
 });
