+++ client/app/src/app/core/export-data/export-data-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { ExportDataComponent } from './export-data/export-data.component';
-import { LoginGuard } from '../auth/_guards/login-guard.service';
+import { LoginGuard } from '../../shared/_guards/login-guard.service';
 import { CreateProfileGuardService } from '../../user-profile/create-profile-guard.service';
 const routes = [{
   path: 'export-data',
