+++ client/app/src/app/tangy-forms/tangy-forms-routing.module.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs'
 import { RouterModule, Routes } from '@angular/router';
 import { CreateProfileGuardService } from '../user-profile/create-profile-guard.service';
 
-import { LoginGuard } from '../core/auth/_guards/login-guard.service';
+import { LoginGuard } from '../shared/_guards/login-guard.service';
 import { TangyFormsPlayerComponent } from './tangy-forms-player/tangy-forms-player.component';
 import { _TRANSLATE } from '../shared/translation-marker';
