+++ client/app/src/app/user-profile/user-profile-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { LoginGuard } from '../core/auth/_guards/login-guard.service';
+import { LoginGuard } from '../shared/_guards/login-guard.service';
 import { UserProfileComponent } from './user-profile.component';
 import { ImportUserProfileComponent } from './import-user-profile/import-user-profile.component';
 import { CreateProfileGuardService } from './create-profile-guard.service'
